tests/wpt/tests/trusted-types/trusted-types-eval-reporting-report-only.html

<!DOCTYPE html>
<head>
  <script nonce="123" src="/resources/testharness.js"></script>
  <script nonce="123" src="/resources/testharnessreport.js"></script>
  <script nonce="123" src="./support/csp-violations.js"></script>
</head>
<body>
  <script nonce="123">
  // CSP insists the "trusted-types: ..." directives are delivered as headers
  // (rather than as "<meta http-equiv" tags). This test assumes the following
  // headers are set in the .headers file:
  //
  //   Content-Security-Policy: script-src 'unsafe-inline' 'unsafe-eval'; report-uri ...
  //   Content-Security-Policy: connect-src 'none'
  //   Content-Security-Policy-Report-Only: require-trusted-types-for 'script'
  //
  // The last rule is there so we can provoke a CSP violation report at will.
  // The intent is that in order to test that a violation has *not* been thrown
  // (and without resorting to abominations like timeouts), we force a *another*
  // CSP violation (by violating the img-src rule) and when that event is
  // processed we can we sure that an earlier event - if it indeed occurred -
  // must have already been processed.

  // A sample policy we use to test trustedTypes.createPolicy behaviour.
  const id = x => x;
  const a_policy = {
    createHTML: id,
    createScriptURL: id,
    createScript: id,
  };

  const scriptyPolicy = trustedTypes.createPolicy('allowEval', a_policy);

  window.script_run_beacon = 'vanilla';

  promise_test(async t => {
    const input = 'script_run_beacon="report-only-does-not-stop"';
    let violation = await trusted_type_violation_without_exception_for(_ =>
      eval(input)
    );
    assert_true(violation.originalPolicy.includes("require-trusted-types-for 'script'"));
    assert_equals(violation.sample, `eval|${clipSampleIfNeeded(input)}`);
    assert_equals(script_run_beacon, 'report-only-does-not-stop');
  }, "Trusted Type violation report: evaluating a string.");

  promise_test(async t => {
    await no_trusted_type_violation_for(_ =>
      eval(scriptyPolicy.createScript('script_run_beacon="trusted-script-ok"'))
    );
    assert_equals(script_run_beacon, 'trusted-script-ok');
  }, "Trusted Type violation report: evaluating a Trusted Script.");

  promise_test(async t => {
    trustedTypes.createPolicy('default', {
      createScript: s => s,
    });
    await no_trusted_type_violation_for(_ =>
      eval('script_run_beacon="payload"')
    );
    assert_equals(script_run_beacon, 'payload');
  }, "Trusted Type violation report: default policy runs in report-only mode.");

  </script>
</body>
